[Build] Reinforce the Gradle configuration for sample folder selection
In the project, a folder walk is done to list all the "samples" folders.
This is done by checking the path in the settings.gradle.kts file.
The path provided in the filter logic can be absolute and in this case it can include parents folders names, which can corrupt the check.

Example:
My project is cloned in the following folder: `<user-home>/android/samples/platform-samples`
Then all the folders will be considered as samples and the Gradle build will failed at configuration step.

Solution:
Remove the parents folders part in the path during the filtering if the path is absolute.