Add Android Studio Previews for the action list layout.
As a developer, I would be most interested to tweak breakpoints and possibly view my widget at standard widget sizes as I develop the layout.

So, using the sizing guidance published on DAC, created custom annotations for different size sets that can be reused across layouts for viewing them at those sizes besides the breakpoints for that layout.

When focusing on a single breakpoint, developer can comment the other preview sizes if needed. But, having the ability to view them all in a grid makes verification quite faster.